(test) replace all "done (_.prettify(error))" with "done.fail (_.prettify(error))"
done (error) should cause the test to fail, but perhaps the passage through _.prettify
is enough for Jasmine not longer to recognise it as an error.

This change caused thirteen tests to fail; I have repaired the one pertinent to my
work and X'ed out and "TODO"ed the remainder, which will be fixed in a separate story.