(fix) test and implement a fix for failure to maintain state properly when processing an update without one.
This concerns the situation when a default state is provided, i.e. all products must have
states, and an update is received for a product the existing version of which has a state,
but the version from the CSV file does not.  This situation arises when both the original
and update CSV files do not specify state, and the state is assigned during the initial
import from the default state.

Before this fix, this resulted in a transitionState action being created without specifying
what state to transition to, which led to the update being rejected with "Request body does
not contain valid JSON."